# Holy Grail Layout
The Holy Grail Layout is a classic CSS problem with various solutions presented over time. If you’re unfamiliar with the history of the Holy Grail layout, this A List Apart article offers a pretty good summary and links to a few of the more well-known solutions.

At its core, the Holy Grail Layout is a page with a header, footer, and three columns. The center column contains the main content, and the left and right columns contain supplemental content like ads or navigation.

Most CSS solutions to this problem aim to meet a few goals:

* They should have a fluid center with fixed-width sidebars.
* The center column (main content) should appear first in the HTML source.
* All columns should be the same height, regardless of which column is actually the tallest.
* They should require minimal markup.
* The footer should “stick” to the bottom of the page when content is sparse.
* Unfortunately, because of the nature of these goals and the original limitations of CSS, none of the classic solutions to this problem were ever able to satisfy all of them.

With Flexbox, a complete solution is finally possible.

## The HTML