## Being Responsive
The Holy Grail layout came from an era of Web design when pretty much everyone was browsing on a computer. But with the increasing number of mobile devices and the rising popularity of responsive design, the Holy Grail layout has gone mostly out of fashion.

Either way, with Flexbox, creating a mobile-first and mobile-friendly version of the Holy Grail layout is easy. The gist is to simply make the center section flex-direction:column by default and then flex-direction:row for larger screens.

Here’s a complete example that is responsive and mobile-first. You can also resize this browser window to see it in action.